requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
django-modeltranslation==0.18.5
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
Pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3
zipp==3.9.0
